function-runner/pyproject.toml
[tool.poetry]
name = "open-interpreter-code-runner"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
open-interpreter = "^0.1.9"
pydantic = "^2.3.0"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
python-json-logger = "^2.0.7"
google-cloud-storage = "^2.11.0"
pandas = "^2.1.1"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
japanize-matplotlib = "^1.1.3"
scikit-learn = "^1.3.1"
mplfinance = "^0.12.10b0"
yfinance = "^0.2.31"
scrapy = "^2.11.0"
beautifulsoup4 = "^4.12.2"
opencv-python = "^4.8.1.78"
ffmpeg-python = "^0.2.0"
pymupdf = "^1.23.5"
pytube = "^15.0.0"
pyocr = "^0.8.5"
easyocr = "^1.7.1"
pydub = "^0.25.1"
pdfkit = "^1.0.0"
weasyprint = "^60.1"
toml = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

